+++ packages/keyring-api/src/eth/index.ts
@@ -1,5 +1,4 @@
 export * from './constants';
-export * from './erc4337';
 export * from './rpc';
 export * from './types';
 export * from './utils';