+++ packages/keyring-internal-api/src/index.ts
@@ -1,3 +1,2 @@
 export * from './compatibility';
-export type * from './eth';
 export * from './types';